Moving all Login stuff into a new route for /v3/auth
