+++ lib/registerPlugins.js
@@ -37,12 +37,11 @@ function registerDefaultPlugins(server, callback) {
  */
 function registerResourcePlugins(server, config, callback) {
     const plugins = [
-        'login',
+        'auth',
         'builds',
         'jobs',
         'pipelines',
         'webhooks',
-        'crumb',
         'stats'
     ];
