+++ lib/server.js
@@ -83,8 +83,10 @@ module.exports = (config, callback) => {
         // Initialize common data in buildFactory
         if (server.app.buildFactory) {
             server.app.buildFactory.apiUri = server.info.uri;
-            server.app.buildFactory.tokenGen = (buildId) =>
-                server.plugins.login.generateToken(buildId, ['build']);
+            server.app.buildFactory.tokenGen = (buildId, metadata) =>
+                server.plugins.auth.generateToken(
+                    server.plugins.auth.generateProfile(buildId, ['build'], metadata)
+                );
         }
 
         // Start the server
