+++ plugins/auth/key.js
@@ -1,0 +1,26 @@
+'use strict';
+
+const schema = require('screwdriver-data-schema');
+
+/**
+ * Display the public key for verifying the JWT
+ * @method key
+ * @param  {Object} options
+ * @param  {String} options.jwtPublicKey  Public Key for verifying a JWT was signed by us
+ * @return {Object}                       Hapi Plugin Route
+ */
+module.exports = (options) => ({
+    method: ['GET'],
+    path: '/auth/key',
+    config: {
+        description: 'jwt public key',
+        notes: 'Public Key for verifying JSON Web Tokens',
+        tags: ['api', 'auth', 'key'],
+        handler: (request, reply) => reply({
+            key: options.jwtPublicKey
+        }),
+        response: {
+            schema: schema.api.auth.key
+        }
+    }
+});
