+++ test/plugins/data/jwt.private.key
@@ -1,0 +1,27 @@
+-----BEGIN RSA PRIVATE KEY-----
+MIIEpAIBAAKCAQEAtn9p4XXo1nulcBYjbMaVuKUrTn5lYTXgRdByWVg6lW97JK8A
++on7NrFV5YgZuU2pwnpm+iIGrtNEcHNIlLhUkG+N0BP7OXORFxrBo0nP+hZsy+Ev
+zWyUIVMqEzljHJ2nA8Ln3jCgjlV2iAV0sh8emio1W/nsBPVoJIcqeNMBIbvbrfHd
+XHd5vrN0ixDxguMgGRe6rTx2wgWciHrDCkHxq4hus1n7b0QEPo1BoDfje1QfKqcM
+u6xhV5YLMlF95uqNJCEgnbpMhqCIu6kqH/6hW6K3lImir0huJlAqf6lXFIB3tKxm
+guTqg4CpObDdr65mzle/ZM17bcmhOn/SeEgWoQIDAQABAoIBAQCfnQqW5ZBSXuxP
+/xzWm00BpHwe989lWUCg7vHEELWUU1+JswIRHjM4H5u31/RfrHxIGtQL+0yLBT47
+d6jIYfucI+UNUxyLn/cLFRy82ubqnUhkfLxZV5irGNiY3b1AnuYfFQ3EuBUEx9yW
+lOb+jtcuELKAkjKKF7URFaLg7/NycygPrMcmUikA1B1WPRtWW4PhDikeGAjV5sOQ
+QYXv7PXTmMc6QF38j4qk3THsMqcxQWYp718F/HMOOwA6C3rFPuTw4+xDZDh4s/8C
+dk5WRr/vonqpIuvTo0MrFU1WfIOMUfSYv9sio46KplAxbD1OoDc6ouia9jvJIaSI
+oDomWbcdAoGBAOB85wA6j2rv3Cki1qTscP6rGQYk5tffuKgY4Y8uQzdObRRC37jk
+YDtoyddkE0xJCWT9Uti4g0VIO2bVO/nIXNzXAlNg92SWuqgJNnW8b+WIWIcGpeW7
+j/9IbdTaQaCgF/JDrcpuJXUw/j55QO6X3Ss4ej7AGwlML+nOPSIB1hYXAoGBANAd
+kgRqv9VgcbIqJkrW3OrKrG6UztbHeCHdyfkU7A5G+npmS99hDDmvm+l/nJsAiRHM
+cMc/wZ4tGVNokWtNPZFufhxFOQL9SkQ6v0Up6DNILsLeqKivrU79XCrDJXy9A3pC
+Xppa2MbMgnwU9XG7H27DvwmJ1n1Mty6F0T5X5+QHAoGAAhtJ86eZBvDfSQeJqhyJ
+5unmHzejbu0rh3BMVZKikNdMkL8b6FlESepu0nWe9L1OuELkTyJUIQdPf2SK3oox
+zOS7odqsF6aOc9s6RpqRFj9UTVgToHdFH6knVrMBoKc/8kcOT3L/KP1ftWhf4p1M
+Nwadpw8YcdrkLDaCPlp0DtsCgYEAxlgoXcVPZxEvytmY/SKiEbNve0SfjNs2hzBV
+2eoyv7it6BGupgpcGG5hbbyNiLyN5kSuB3egFBxvUqsKTPWkhAKGdnMSzuKP8U78
+E/LItndI+bgzjfFAf1PAj5DFHYkW02KFvx5uE6m9gnyHIAly/6HXZEPWBD0lKFWp
+An3NJ+kCgYBT4CCvPQ8i1VHVnqtkKVDOO5poKL51zf0FMIG8Rmjk4+PUxWH6aRjd
+JRO9UBRGfWEFUn8vj4PxhXHiAFagryle38YhcM/DVNOxOvjCqrFK8DD68V0RGgen
+oaQ6utfChQZDDSYl6d8sXKKQeU/0mfoKIsse+Am7IHbkW5rmmLvHLA==
+-----END RSA PRIVATE KEY-----
