+++ test/plugins/data/jwt.public.key
@@ -1,0 +1,9 @@
+-----BEGIN PUBLIC KEY-----
+MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAtn9p4XXo1nulcBYjbMaV
+uKUrTn5lYTXgRdByWVg6lW97JK8A+on7NrFV5YgZuU2pwnpm+iIGrtNEcHNIlLhU
+kG+N0BP7OXORFxrBo0nP+hZsy+EvzWyUIVMqEzljHJ2nA8Ln3jCgjlV2iAV0sh8e
+mio1W/nsBPVoJIcqeNMBIbvbrfHdXHd5vrN0ixDxguMgGRe6rTx2wgWciHrDCkHx
+q4hus1n7b0QEPo1BoDfje1QfKqcMu6xhV5YLMlF95uqNJCEgnbpMhqCIu6kqH/6h
+W6K3lImir0huJlAqf6lXFIB3tKxmguTqg4CpObDdr65mzle/ZM17bcmhOn/SeEgW
+oQIDAQAB
+-----END PUBLIC KEY-----
